Rename helper to harness; set some funcs to private; turn init into constructor
